added games to order + db stored procedure ophalen

+++ WebShop-S2/WebShop-S2/Models/ShoppingListViewModel.cs
@@ -7,6 +7,6 @@ namespace WebShop_S2.Models
     public class ShoppingListViewModel
     {
         public List<Game> ShoppingList { get; set; }
-        public Double TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
     }
 }
